+++ src/solve/brute_force_search.cc
@@ -11,7 +11,7 @@ BruteForceSearch::~BruteForceSearch() {
 }
 
 void BruteForceSearch::SearchPermutations(
-  vector<int> current_path, int current_distance,
+  vector<int> current_path, double current_distance,
   unordered_set<int> unvisited_nodes) {
   if (unvisited_nodes.size() == 0) {
     if (current_path.size() == graph()->num_nodes()) {
@@ -28,14 +28,14 @@ void BruteForceSearch::SearchPermutations(
        itr != unvisited_nodes.end();
        ++itr) {
     int node_j = *itr;
-    int weight = 0;
+    double weight = 0;
     if (current_path.size() > 0) {
       weight = graph()->GetEdgeWeight(current_path.back(), node_j);
     }
     if (current_path.size() == 0 || weight) {
       vector<int> new_current_path = current_path;
       new_current_path.push_back(node_j);
-      int new_current_distance = current_distance + weight;
+      double new_current_distance = current_distance + weight;
       unordered_set<int> new_unvisited_nodes = unvisited_nodes;
       new_unvisited_nodes.erase(new_unvisited_nodes.find(node_j));
       SearchPermutations(new_current_path, new_current_distance, new_unvisited_nodes);
@@ -47,7 +47,7 @@ Solution BruteForceSearch::ComputeSolution(const Graph* graph) {
   Reset(graph);
   vector<int> current_path;
   current_path.push_back(0);
-  int current_distance = 0;
+  double current_distance = 0;
   unordered_set<int> unvisited_nodes;
   for (unsigned int i = 1; i < this->graph()->num_nodes(); ++i) {
     unvisited_nodes.insert(i);
